requirements.txt
# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Excel Processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1  # Para arquivos .xls antigos

# Word Processing
python-docx>=1.1.0

# Archive Processing
rarfile>=4.1  # Para suporte a arquivos RAR

# Utils
python-dateutil>=2.8.0
num2words>=0.5.13  # Para valor por extenso

# Production Server
gunicorn>=21.0.0

# Development
python-dotenv>=1.0.0
